fix SMB pivot issue
there was a 32bit field being sent as a 64bit number

+++ teamserver/pkg/agent/agent.go
@@ -587,7 +587,7 @@ func (a *Agent) PivotAddJob(job Job) {
 		Command: COMMAND_PIVOT,
 		Data: []interface{}{
 			DEMON_PIVOT_SMB_COMMAND,
-			AgentID,
+			uint32(AgentID),
 			Packer.Buffer(),
 		},
 	}
@@ -617,7 +617,7 @@ func (a *Agent) PivotAddJob(job Job) {
 			Command: COMMAND_PIVOT,
 			Data: []interface{}{
 				DEMON_PIVOT_SMB_COMMAND,
-				AgentID,
+				uint32(AgentID),
 				Packer.Buffer(),
 			},
 		}
